pyproject.toml
[tool.poetry]
name = "pedalo"
version = "0.1.0"
description = ""
authors = ["Luke31 <lukas.m.shcmid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.258"
streamlit = "^1.25.0"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
tabulate = "^0.9.0"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
neo4j = "^5.11.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
